// apps/admin-web/src/.generated/api/model/updateSystemSettingsRequestDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { DeliveryConfigDto } from './deliveryConfigDto';
import type { OAuthConfigDto } from './oAuthConfigDto';
import type { UpdateAdminEmailConfigRequestDto } from './updateAdminEmailConfigRequestDto';
import type { WebhookConfigDto } from './webhookConfigDto';

export interface UpdateSystemSettingsRequestDto {
  webhook?: WebhookConfigDto;
  delivery?: DeliveryConfigDto;
  oauth?: OAuthConfigDto;
  adminEmail?: UpdateAdminEmailConfigRequestDto;
}